# Speed up `print_annotations` 