--- a/process_frames.py
+++ b/process_frames.py
@@ -13,7 +13,7 @@ from image_diff import calculate_image_difference, calculate_image_hash_differen
 from openai_api import OpenAI_API
 import time
 from thread_safe import shared_data_put_data, shared_data_put_line, ThreadSafeData
-from PIL import Image, ImageDraw, ImageFont
+from PIL import Image, ImageDraw, ImageFilter
 from pathlib import Path
 import pickle
 import imagehash
@@ -58,6 +58,7 @@ class FrameProcessor:
         #print(self.dialogues)
         #TODO remove this from this class and store this somewhere else, so its multi user
         self.previous_image = Image.new('RGB', (100, 100), (255, 255, 255))
+        self.background_image = Image.new('RGB', (100, 100), (255, 255, 255))
         self.last_played = -1 #TODO this should be per user
 
         
@@ -313,6 +314,23 @@ class FrameProcessor:
             return str, result
         else:
             return '', {}
+        
+    def _generate_blurred_image(self, pil_image):
+        # Blurring the background text
+        blurred_image = pil_image.filter(ImageFilter.GaussianBlur(10))
+
+        mask = Image.new("L", pil_image.size, 0)
+
+        draw_mask = ImageDraw.Draw(mask)
+
+        for res_ in self.last_annotations:
+            # Draw the rectangle on the mask
+            draw_mask.rectangle(res_[0], fill=255)
+
+        image_with_blur = Image.composite(blurred_image, pil_image, mask)
+
+        return image_with_blur
+
 
     def process_frame(self, frame_pil, frame_count, fps):
         """
@@ -352,8 +370,9 @@ class FrameProcessor:
 
                 self.previous_image = img
                 self.last_annotations = annotations
+                self.background_image = self._generate_blurred_image(img)
                 
-                return last_played, self.previous_image, highlighted_image, annotations, translation
+                return last_played, self.previous_image, highlighted_image, annotations, translation, self.background_image
 
             # crop the image to top half
             img_crop = image_crop_in_top_half(img)
@@ -424,12 +443,13 @@ class FrameProcessor:
 
             self.previous_image = img
             self.last_annotations = annotations
+            self.background_image = self._generate_blurred_image(img)
             
-            return last_played, self.previous_image, highlighted_image, annotations, translation
+            return last_played, self.previous_image, highlighted_image, annotations, translation, self.background_image
         else:
             print("Difference is less than 10%. No need to call OCR again.")
             
-            return None, None, None, self.last_annotations, None
+            return None, None, None, self.last_annotations, None, self.background_image
         
     def create_output_dirs(self):